pyproject.toml
[project]
name = "pikapikagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.35.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "sentence-transformers>=5.0.0",
    "torchvision>=0.22.1",
    "transformers>=4.53.0",
]
